+++ common/config/rush/repo-state.json
@@ -1,5 +1,5 @@
 // DO NOT MODIFY THIS FILE MANUALLY BUT DO COMMIT IT. It is generated and used by Rush.
 {
-  "pnpmShrinkwrapHash": "cd56000d9fdb9fad5e1d292089f000cadca98f13",
+  "pnpmShrinkwrapHash": "8c7d02dc2b2ad6a028fdaf27256044455ace3aa2",
   "preferredVersionsHash": "bf21a9e8fbc5a3846fb05b4fa0859e0917b2202f"
 }
