+++ packages/libs/kadena.js/package.json
@@ -57,6 +57,7 @@
   },
   "dependencies": {
     "@kadena/cryptography-utils": "workspace:*",
+    "@kadena/pactjs-utils": "workspace:*",
     "@kadena/types": "workspace:*",
     "bignumber.js": "^9.0.2",
     "blakejs": "^1.2.1",
