+++ packages/libs/pactjs-utils/.eslintrc.js
@@ -1,0 +1,7 @@
+// This is a workaround for https://github.com/eslint/eslint/issues/3458
+require('@rushstack/eslint-config/patch/modern-module-resolution');
+
+module.exports = {
+  extends: ['@rushstack/eslint-config/profile/node'],
+  parserOptions: { tsconfigRootDir: __dirname },
+};
