+++ packages/libs/pactjs-utils/README.md
@@ -1,0 +1,20 @@
+<p align="center">
+  <picture>
+    <source srcset="../../../common/images/Kadena.JS_logo-white.png" media="(prefers-color-scheme: dark)"/>
+    <img src="../../../common/images/Kadena.JS_logo-black.png" width="200" alt="kadena.js logo" />
+  </picture>
+</p>
+
+# kadena.js - Pactjs-Utils
+
+> Pactjs-Utils is a collection of utility functions pactjs related
+
+API Reference can be found here
+[./etc/pactjs-utils.api.md](./etc/pactjs-utils.api.md)
+
+<hr>
+
+crypto:
+
+- `createExp`
+- `PactNumber`
