+++ packages/libs/pactjs-utils/config/rig.json
@@ -1,0 +1,5 @@
+{
+  "$schema": "https://developer.microsoft.com/json-schemas/rig-package/rig.schema.json",
+
+  "rigPackageName": "@kadena/heft-rig"
+}
