+++ packages/libs/pactjs-utils/etc/pactjs-utils.api.md
@@ -1,0 +1,34 @@
+## API Report File for "@kadena/pactjs-utils"
+
+> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
+
+```ts
+
+import { IPactDecimal } from '@kadena/types';
+import { IPactInt } from '@kadena/types';
+import { PactCode } from '@kadena/types';
+import { PactValue } from '@kadena/types';
+
+// @alpha
+export function createExp(firstArg: string, ...args: PactValue[]): PactCode;
+
+// @alpha
+export class PactNumber {
+    constructor(value: string | number);
+    // (undocumented)
+    toDecimal(): string;
+    // (undocumented)
+    toInteger(): string;
+    // (undocumented)
+    toPactDecimal(): IPactDecimal;
+    // (undocumented)
+    toPactInteger(): IPactInt;
+    // (undocumented)
+    toStringifiedDecimal(): string;
+    // (undocumented)
+    toStringifiedInteger(): string;
+}
+
+// (No @packageDocumentation comment for this package)
+
+```
