+++ packages/libs/pactjs-utils/tsconfig.json
@@ -1,0 +1,6 @@
+{
+  "extends": "./node_modules/@kadena/heft-rig/profiles/default/tsconfig-base.json",
+  "compilerOptions": {
+    "types": ["heft-jest", "node"]
+  }
+}
