+++ rush.json
@@ -428,6 +428,13 @@
     },
 
     {
+      "packageName": "@kadena/pactjs-utils",
+      "projectFolder": "packages/libs/pactjs-utils",
+      "tags": ["libs"],
+      "shouldPublish": false
+    },
+
+    {
       "packageName": "@kadena/integration-tests",
       "projectFolder": "packages/tools/integration-tests",
       "tags": ["tools"],
